refactor(control): split cache persistence
